Refactor: viewController split up
