+++ JuiceMaker/JuiceMaker/Controller/ViewController.swift
@@ -46,105 +46,14 @@ class ViewController: UIViewController {
 		changeTextOn(label: kiwiStockLabel, about: .kiwi)
 		changeTextOn(label: mangoStockLabel, about: .mango)
 		changeTextOn(label: pineappleStockLabel, about: .pineapple)
-		
-		guard let strawberryStepper = strawberryStepper,
-			  let strawberryStore = findFruitStore(about: Fruit.strawberry),
-			  let bananaStepper = bananaStepper,
-			  let bananaStore = findFruitStore(about: Fruit.banana),
-			  let pineappleStepper = pineappleStepper,
-			  let pineappleStore = findFruitStore(about: Fruit.pineapple),
-			  let kiwiStepper = kiwiStepper,
-			  let kiwiStore = findFruitStore(about: Fruit.kiwi),
-			  let mangoStepper = mangoStepper,
-			  let mangoStore = findFruitStore(about: Fruit.mango)
-		else {
-			return
-		}
-		
-		strawberryStepper.value = Double( strawberryStore.stock )
-		bananaStepper.value = Double( bananaStore.stock )
-		pineappleStepper.value = Double( pineappleStore.stock )
-		kiwiStepper.value = Double( kiwiStore.stock )
-		mangoStepper.value = Double( mangoStore.stock )
-		self.navigationItem.title = "재고수정"
-		self.navigationItem.hidesBackButton = true
 	}
-	
-	@IBAction func exitFuirtStore(_ sender: UIButton) {
-		self.navigationController?.popViewController(animated: true)
-	}
-	
-	func findFruitStore(about: Fruit) -> FruitStore? {
-		return ViewController
-			.juiceMaker
-			.fruitStores
-			.filter({ store in
-				store.name == about
-			}).first
-	}
-	
-	@IBAction func changedStrawberryStock(_ sender: UIStepper) {
-		guard let store = findFruitStore(about: Fruit.strawberry) else {
-			return
-		}
-		
-		store.setStock(amount: Int(sender.value) - store.stock )
-		
-		changeTextOn(label: strawberryStockLabel, about: .strawberry)
-	}
-	
-	@IBAction func changedPineappleStock(_ sender: UIStepper) {
-		guard let store = findFruitStore(about: Fruit.pineapple) else {
-			return
-		}
-		
-		store.setStock(amount: Int(sender.value) - store.stock )
-		
-		changeTextOn(label: pineappleStockLabel, about: .pineapple)
-	}
-	
-	
-	@IBAction func changedMangoStock(_ sender: UIStepper) {
-		guard let store = findFruitStore(about: Fruit.mango) else {
-			return
-		}
-		
-		store.setStock(amount: Int(sender.value) - store.stock )
-		
-		changeTextOn(label: mangoStockLabel, about: .mango)
-	}
-	
-	@IBAction func changedKiwiStock(_ sender: UIStepper) {
-		guard let store = findFruitStore(about: Fruit.kiwi) else {
-			return
-		}
-		
-		store.setStock(amount: Int(sender.value) - store.stock )
-		
-		changeTextOn(label: kiwiStockLabel, about: .kiwi)
-	}
-	
-	@IBAction func changedBananaStock(_ sender: UIStepper) {
-		guard let store = findFruitStore(about: Fruit.banana) else {
-			return
-		}
-		
-		store.setStock(amount: Int(sender.value) - store.stock )
-		
-		changeTextOn(label: bananaStockLabel, about: .banana)
-	}
+
 	
 	@IBOutlet weak var strawberryStockLabel: UILabel!
 	@IBOutlet weak var bananaStockLabel: UILabel!
 	@IBOutlet weak var pineappleStockLabel: UILabel!
 	@IBOutlet weak var kiwiStockLabel: UILabel!
 	@IBOutlet weak var mangoStockLabel: UILabel!
-	
-	@IBOutlet weak var strawberryStepper: UIStepper!
-	@IBOutlet weak var bananaStepper: UIStepper!
-	@IBOutlet weak var pineappleStepper: UIStepper!
-	@IBOutlet weak var kiwiStepper: UIStepper!
-	@IBOutlet weak var mangoStepper: UIStepper!
 	
 	@IBAction func makeStrawNanaJuice(_ sender: UIButton) {
 		let currentMenu: JuiceMenu = .strawNanaJuice
